+++ packages/expo-video/ios/VideoPlayer.swift
@@ -6,7 +6,7 @@ import ExpoModulesCore
 
 internal final class VideoPlayer: SharedRef<AVPlayer>, Hashable, VideoPlayerObserverDelegate {
   lazy var contentKeyManager = ContentKeyManager()
-  var observer: VideoPlayerObserver?
+  let observer: VideoPlayerObserver
 
   var loop = false
   private(set) var isPlaying = false
@@ -62,16 +62,30 @@ internal final class VideoPlayer: SharedRef<AVPlayer>, Hashable, VideoPlayerObse
     }
   }
 
+  var showNowPlayingNotification = false {
+    didSet {
+      // The audio session needs to be appropriate before displaying the notfication
+      VideoManager.shared.setAppropriateAudioSessionOrWarn()
+
+      if showNowPlayingNotification {
+        NowPlayingManager.shared.registerPlayer(self)
+      } else {
+        NowPlayingManager.shared.unregisterPlayer(self)
+      }
+    }
+  }
+
   override init(_ pointer: AVPlayer) {
+    observer = VideoPlayerObserver(player: pointer)
     super.init(pointer)
-    observer = VideoPlayerObserver(player: pointer, delegate: self)
-    NowPlayingManager.shared.registerPlayer(pointer)
+    observer.registerDelegate(delegate: self)
     VideoManager.shared.register(videoPlayer: self)
   }
 
   deinit {
-    NowPlayingManager.shared.unregisterPlayer(pointer)
+    NowPlayingManager.shared.unregisterPlayer(self)
     VideoManager.shared.unregister(videoPlayer: self)
+    observer.unregisterDelegate(delegate: self)
     pointer.replaceCurrentItem(with: nil)
   }